Do not consider a container deletion to conflict with unsets under it
Since a containment deletion and the "unset"(s) under it were seen as a
pseudo conflict with each other, we were "hiding" the deletion from view
and preventing the merge altogether.

This adds an explicit dependency between the core tests and objenesis to
avoid failures when launching the build locally (mockito has a transient
dependency to objenesis).

+++ plugins/org.eclipse.emf.compare.tests/src/org/eclipse/emf/compare/tests/conflict/PseudoConflictDetectionTest.java
@@ -55,7 +55,7 @@ public class PseudoConflictDetectionTest {
 		 */
 		Object[] test1 = new Object[] {inputData.getPseudoConflictCase1Ancestor(),
 				inputData.getPseudoConflictCase1RightAndLeftModel(),
-				inputData.getPseudoConflictCase1RightAndLeftModel(), new Integer(4) };
+				inputData.getPseudoConflictCase1RightAndLeftModel(), new Integer(5) };
 		/**
 		 * Right and left models add the same univaluated containment EReference.
 		 */
